Moved violins to analyse script
