+++ src/analyse.py
@@ -8,7 +8,8 @@ from pathlib import Path
 from data.util.subgroup import get_healthy, normalise_bps
 from features.descriptive import demographics, flowchart
 from features.comparative import smoking
-from models.regression import univariate, multivariate
+from models.linear import univariate, multivariate
+from visualization import violin
 
 runs = ["descriptive", "comparative", "regression", "clustering", "visualisation"]
 
@@ -53,7 +54,7 @@ def main(args):
             multivariate.fit_analyse(data, bps, out_path, True, logger) 
         ),
         runs[3]: lambda: (),
-        runs[4]: lambda: (),
+        runs[4]: lambda: (violin.make_plots(data, bps, out_path)),
     }
 
     # Run the analysis based on the desired run
