+++ src/models/linear/univariate.py
